+++ format/format_suite_test.go
@@ -1,0 +1,13 @@
+package format_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestFiles(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Formatting Suite")
+}